# dict/models.ts
/* eslint-disable */
// tslint:disable
/**
 * ServiceApi 1-
 * 
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

export namespace Api {
	/**
	 * <p>ResponseInStatusCode</p>
	 */
	export interface ApiResponse {
		/**
		 * @type {number}
		 * @memberof ApiResponse
		 */
		code?: number;
		type?: string;
		message?: string;
	}

	export interface BodyData1 {
		First_Name?: string;
		Middle_Name?: string;
		Last_Name?: string;
		/**
		 * @type {number}
		 * @memberof BodyData1
		 */
		DateOfBirth?: number;
		Occupation?: string;
		Visa_Type?: string;
		/**
		 * @type {number}
		 * @memberof BodyData1
		 */
		Account_Number?: number;
		SSN?: string;
	}

	export interface BodyData2 {
		First_Name?: string;
		Middle_Name?: string;
		Last_Name?: string;
		/**
		 * @type {number}
		 * @memberof BodyData2
		 */
		DateOfBirth?: number;
		Occupation?: string;
		Visa_Type?: string;
		/**
		 * @type {number}
		 * @memberof BodyData2
		 */
		Account_Number?: number;
		/**
		 * @type {number}
		 * @memberof BodyData2
		 */
		Passport_Number?: number;
		Passport_Expiry_Date?: string;
		/**
		 * @type {number}
		 * @memberof BodyData2
		 */
		Visa_Number?: number;
		Visa_Expiry_Date?: string;
		First_Entry_Date_into_USA?: string;
	}

}